catalyst: Set makeopts via catalyst.conf
We currently have two mechanisms of setting MAKEOPTS: in spec files and
in catalystrc.

Setting makeopts in spec files doesn't make sense. The spec should
describe the thing that's being built and not contain options that are
specific to the build system.

Setting makeopts via catalystrc is better, but it only applies to the
actual build system invocations, leaving emerge to run jobs serially or
again requiring configuration specific to the build machine to be put
into the spec file. For example:

	update_seed_command: ... --jobs 5 --load-average 5

With makeopts specified in catalyst.conf, catalyst has the information
required to configure both emerge and the build systems emerge executes.

This removes the undocumented makeopts spec file option and replaces it
with a setting in catalyst.conf.

Signed-off-by: Matt Turner <mattst88@gentoo.org>